backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0
alembic>=1.13

# PDF parsing
pdfplumber>=0.11

# Job scraping
python-jobspy>=1.1.70
playwright>=1.45

# ML / Matching
sentence-transformers>=3.0
scikit-learn>=1.5
numpy>=1.26

# Utilities
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0
httpx>=0.27
fake-useragent>=1.5

# Testing
pytest>=8.0
pytest-asyncio>=0.24
